refactor: extract async locate/sound methods to LocateOperations
Extract Passthrough API methods from main.py to locate.py:
- async_locate_device (~275 lines)
- async_play_sound (~55 lines)
- async_stop_sound (~70 lines)

main.py: 4,961 → 4,561 lines (-400)
locate.py: 202 → 629 lines (+427)
